// src/types.ts
export interface Topic {
    id: string;
    name: string;
}

export interface Subscription {
    id: string;
    name: string;
}